backend/requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
sqlalchemy==2.0.*
pymysql==1.1.*
pydantic==2.*
cryptography
google-generativeai
python-dotenv
